⭐ FEAT : NFT 구매, 판매, 거래 내역 리스트 기능 추가

See merge request s08-blockchain-nft-sub2/S08P22A307!21

+++ backend/nfast/src/main/java/com/nft/nfast/model/dto/user/TradeListDto.java
@@ -18,7 +18,7 @@ public class TradeListDto {
     private BigDecimal tradeListPrice;
     private Date tradeListDate;
     private Byte tradeListType;
-    private Byte userSequence;
+    private long userSequence;
 
     public TradeList toEntity(){
         TradeList tradeList = TradeList.builder()
